Migrate zypper refs to new command format
