+++ src/commands/services.h
@@ -2,5 +2,6 @@
 #define ZYPPER_COMMANDS_SERVICES_H_INCLUDED
 
 #include "services/list.h"
+#include "services/refresh.h"
 
 #endif
